requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.1
dj-database-url==1.0.0
Django==4.1
gunicorn==20.1.0
idna==3.3
psycopg2-binary==2.9.5
# psycopg2==2.9.5
pytz==2022.2.1
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
typed-ast==1.5.4
urllib3==1.26.11
whitenoise==6.2.0
wrapt==1.14.1
django-browser-reload==1.6.0
django-tailwind==3.4.0
djangorestframework==3.14.0
Markdown==3.4.1
django-filter==22.1
python-dotenv==0.21.0